+++ challenge-catar Watch App/Components/TextView/TextView.swift
@@ -1,0 +1,8 @@
+//
+//  TextView.swift
+//  challenge-catar Watch App
+//
+//  Created by alexdamascena on 27/11/22.
+//
+
+import SwiftUI
